latest

+++ mobile/lib/features/home/presentation/home_screen.dart
@@ -211,7 +211,7 @@ class _PendingFeeCard extends StatelessWidget {
   Widget build(BuildContext context) {
     final l = AppLocalizations.of(context)!;
     final amount = (fee['amount'] as num).toDouble();
-    final days = fee['days_until_due'] as int? ?? 0;
+    final days = fee['days_until_due'] as double? ?? 0;
     final isUrgent = days <= 3;
 
     return Container(
@@ -241,7 +241,7 @@ class _PendingFeeCard extends StatelessWidget {
           Text('${amount.toStringAsFixed(0)} ج.م',
               style: const TextStyle(fontWeight: FontWeight.w700, fontSize: 15)),
           Text(
-            days <= 0 ? l.dueToday : l.daysLeft(days),
+            days <= 0 ? l.dueToday : l.daysLeft(days.toInt()),
             style: TextStyle(color: isUrgent ? AppTheme.warning : Colors.grey, fontSize: 11),
           ),
         ]),
